requirements.txt
Flask==2.2.5
Requests==2.31.0
waitress==2.1.2
schedule==1.2.0
python-dotenv==1.0.0
